feat: migrate LoginScreen to :feature:login KMP module
- Create :feature:login KMP module with androidTarget + iOS targets
- Move LoginScreenStateless, LoginUiState, PasswordValidation, all login
  composables (SignInContent, SignUpContent, GoogleSignInButton) to commonMain
- Delete old LoginComponents.kt from :app
- App keeps stateful wrapper with Google Sign-In / CredentialManager / LocalContext
- Add :feature:login to settings.gradle.kts and :app dependencies
- Update app/build.gradle.kts with all feature module dependencies
