+++ settings.gradle.kts
@@ -21,6 +21,7 @@ dependencyResolutionManagement {
 
 rootProject.name = "Yambol"
 include(":app")
+include(":composeApp")
 
 // Phase 0 skeleton modules — code will be migrated here in Phase 1+
 include(":core:domain")
@@ -34,3 +35,4 @@ include(":feature:createTeam")
 include(":feature:playerCard")
 include(":feature:statistics")
 include(":feature:profile")
+include(":feature:login")
